+++ ITMLib/Trackers/CPU/libviso2/matrix.cpp
@@ -27,9 +27,9 @@ using namespace std;
 
 #define SWAP(a,b) {temp=a;a=b;b=temp;}
 #define SIGN(a,b) ((b) >= 0.0 ? fabs(a) : -fabs(a))
-static FLOAT sqrarg;
+static DOUBLE sqrarg;
 #define SQR(a) ((sqrarg=(a)) == 0.0 ? 0.0 : sqrarg*sqrarg)
-static FLOAT maxarg1,maxarg2;
+static DOUBLE maxarg1,maxarg2;
 #define FMAX(a,b) (maxarg1=(a),maxarg2=(b),(maxarg1) > (maxarg2) ? (maxarg1) : (maxarg2))
 static int32_t iminarg1,iminarg2;
 #define IMIN(a,b) (iminarg1=(a),iminarg2=(b),(iminarg1) < (iminarg2) ? (iminarg1) : (iminarg2))
@@ -47,7 +47,7 @@ Matrix::Matrix (const int32_t m_,const int32_t n_) {
   allocateMemory(m_,n_);
 }
 
-Matrix::Matrix (const int32_t m_,const int32_t n_,const FLOAT* val_) {
+Matrix::Matrix (const int32_t m_,const int32_t n_,const DOUBLE* val_) {
   allocateMemory(m_,n_);
   int32_t k=0;
   for (int32_t i=0; i<m_; i++)
@@ -58,7 +58,7 @@ Matrix::Matrix (const int32_t m_,const int32_t n_,const FLOAT* val_) {
 Matrix::Matrix (const Matrix &M) {
   allocateMemory(M.m,M.n);
   for (int32_t i=0; i<M.m; i++)
-    memcpy(val[i],M.val[i],M.n*sizeof(FLOAT));
+    memcpy(val[i],M.val[i],M.n*sizeof(DOUBLE));
 }
 
 Matrix::~Matrix () {
@@ -73,12 +73,12 @@ Matrix& Matrix::operator= (const Matrix &M) {
     }
     if (M.n>0)
       for (int32_t i=0; i<M.m; i++)
-        memcpy(val[i],M.val[i],M.n*sizeof(FLOAT));
+        memcpy(val[i],M.val[i],M.n*sizeof(DOUBLE));
   }
   return *this;
 }
 
-void Matrix::getData(FLOAT* val_,int32_t i1,int32_t j1,int32_t i2,int32_t j2) {
+void Matrix::getData(DOUBLE* val_,int32_t i1,int32_t j1,int32_t i2,int32_t j2) {
   if (i2==-1) i2 = m-1;
   if (j2==-1) j2 = n-1;
   int32_t k=0;
@@ -115,7 +115,7 @@ void Matrix::setMat(const Matrix &M,const int32_t i1,const int32_t j1) {
       val[i1+i][j1+j] = M.val[i][j];
 }
 
-void Matrix::setVal(FLOAT s,int32_t i1,int32_t j1,int32_t i2,int32_t j2) {
+void Matrix::setVal(DOUBLE s,int32_t i1,int32_t j1,int32_t i2,int32_t j2) {
   if (i2==-1) i2 = m-1;
   if (j2==-1) j2 = n-1;
   if (i2<i1 || j2<j1) {
@@ -127,7 +127,7 @@ void Matrix::setVal(FLOAT s,int32_t i1,int32_t j1,int32_t i2,int32_t j2) {
       val[i][j] = s;
 }
 
-void Matrix::setDiag(FLOAT s,int32_t i1,int32_t i2) {
+void Matrix::setDiag(DOUBLE s,int32_t i1,int32_t i2) {
   if (i2==-1) i2 = min(m-1,n-1);
   for (int32_t i=i1; i<=i2; i++)
     val[i][i] = s;
@@ -194,9 +194,9 @@ Matrix Matrix::reshape(const Matrix &M,int32_t m_,int32_t n_) {
   return M2;
 }
 
-Matrix Matrix::rotMatX (const FLOAT &angle) {
-  FLOAT s = sin(angle);
-  FLOAT c = cos(angle);
+Matrix Matrix::rotMatX (const DOUBLE &angle) {
+  DOUBLE s = sin(angle);
+  DOUBLE c = cos(angle);
   Matrix R(3,3);
   R.val[0][0] = +1;
   R.val[1][1] = +c;
@@ -206,9 +206,9 @@ Matrix Matrix::rotMatX (const FLOAT &angle) {
   return R;
 }
 
-Matrix Matrix::rotMatY (const FLOAT &angle) {
-  FLOAT s = sin(angle);
-  FLOAT c = cos(angle);
+Matrix Matrix::rotMatY (const DOUBLE &angle) {
+  DOUBLE s = sin(angle);
+  DOUBLE c = cos(angle);
   Matrix R(3,3);
   R.val[0][0] = +c;
   R.val[0][2] = +s;
@@ -218,9 +218,9 @@ Matrix Matrix::rotMatY (const FLOAT &angle) {
   return R;
 }
 
-Matrix Matrix::rotMatZ (const FLOAT &angle) {
-  FLOAT s = sin(angle);
-  FLOAT c = cos(angle);
+Matrix Matrix::rotMatZ (const DOUBLE &angle) {
+  DOUBLE s = sin(angle);
+  DOUBLE c = cos(angle);
   Matrix R(3,3);
   R.val[0][0] = +c;
   R.val[0][1] = -s;
@@ -276,7 +276,7 @@ Matrix Matrix::operator* (const Matrix &M) {
   return C;
 }
 
-Matrix Matrix::operator* (const FLOAT &s) {
+Matrix Matrix::operator* (const DOUBLE &s) {
   Matrix C(m,n);
   for (int32_t i=0; i<m; i++)
     for (int32_t j=0; j<n; j++)
@@ -319,7 +319,7 @@ Matrix Matrix::operator/ (const Matrix &M) {
   } 
 }
 
-Matrix Matrix::operator/ (const FLOAT &s) {
+Matrix Matrix::operator/ (const DOUBLE &s) {
   if (fabs(s)<1e-20) {
     cerr << "ERROR: Trying to divide by zero!" << endl;
     exit(0);
@@ -347,20 +347,20 @@ Matrix Matrix::operator~ () {
   return C;
 }
 
-FLOAT Matrix::l2norm () {
-  FLOAT norm = 0;
+DOUBLE Matrix::l2norm () {
+  DOUBLE norm = 0;
   for (int32_t i=0; i<m; i++)
     for (int32_t j=0; j<n; j++)
       norm += val[i][j]*val[i][j];
   return sqrt(norm);
 }
 
-FLOAT Matrix::mean () {
-  FLOAT mean = 0;
+DOUBLE Matrix::mean () {
+  DOUBLE mean = 0;
   for (int32_t i=0; i<m; i++)
     for (int32_t j=0; j<n; j++)
       mean += val[i][j];
-  return mean/(FLOAT)(m*n);
+  return mean/(DOUBLE)(m*n);
 }
 
 Matrix Matrix::cross (const Matrix &a, const Matrix &b) {
@@ -397,7 +397,7 @@ bool Matrix::inv () {
   return true;
 }
 
-FLOAT Matrix::det () {
+DOUBLE Matrix::det () {
   
   if (m != n) {
     cerr << "ERROR: Trying to compute determinant of a matrix of size (" << m << "x" << n << ")" << endl;
@@ -406,7 +406,7 @@ FLOAT Matrix::det () {
     
   Matrix A(*this);
   int32_t *idx = (int32_t*)malloc(m*sizeof(int32_t));
-  FLOAT d = 1;
+  DOUBLE d = 1;
   A.lu(idx,d);
   for( int32_t i=0; i<m; i++)
     d *= A.val[i][i];
@@ -414,7 +414,7 @@ FLOAT Matrix::det () {
   return d;
 }
 
-bool Matrix::solve (const Matrix &M, FLOAT eps) {
+bool Matrix::solve (const Matrix &M, DOUBLE eps) {
   
   // substitutes
   const Matrix &A = M;
@@ -433,7 +433,7 @@ bool Matrix::solve (const Matrix &M, FLOAT eps) {
   
   // loop variables
   int32_t i, icol, irow, j, k, l, ll;
-  FLOAT big, dum, pivinv, temp;
+  DOUBLE big, dum, pivinv, temp;
   
   // initialize pivots to zero
   for (j=0;j<m;j++) ipiv[j]=0;
@@ -511,7 +511,7 @@ bool Matrix::solve (const Matrix &M, FLOAT eps) {
 // or odd, respectively. This routine is used in combination with lubksb to solve linear equations
 // or invert a matrix.
 
-bool Matrix::lu(int32_t *idx, FLOAT &d, FLOAT eps) {
+bool Matrix::lu(int32_t *idx, DOUBLE &d, DOUBLE eps) {
   
   if (m != n) {
     cerr << "ERROR: Trying to LU decompose a matrix of size (" << m << "x" << n << ")" << endl;
@@ -519,8 +519,8 @@ bool Matrix::lu(int32_t *idx, FLOAT &d, FLOAT eps) {
   }
   
   int32_t i,imax,j,k;
-  FLOAT   big,dum,sum,temp;
-  FLOAT* vv = (FLOAT*)malloc(n*sizeof(FLOAT)); // vv stores the implicit scaling of each row.
+  DOUBLE   big,dum,sum,temp;
+  DOUBLE* vv = (DOUBLE*)malloc(n*sizeof(DOUBLE)); // vv stores the implicit scaling of each row.
   d = 1.0;
   for (i=0; i<n; i++) { // Loop over rows to get the implicit scaling information.
     big = 0.0;
@@ -582,11 +582,11 @@ void Matrix::svd(Matrix &U2,Matrix &W,Matrix &V) {
   U2 = Matrix(m,m);
   V  = Matrix(n,n);
 
-  FLOAT* w   = (FLOAT*)malloc(n*sizeof(FLOAT));
-  FLOAT* rv1 = (FLOAT*)malloc(n*sizeof(FLOAT));
+  DOUBLE* w   = (DOUBLE*)malloc(n*sizeof(DOUBLE));
+  DOUBLE* rv1 = (DOUBLE*)malloc(n*sizeof(DOUBLE));
 
   int32_t flag,i,its,j,jj,k,l,nm;
-  FLOAT   anorm,c,f,g,h,s,scale,x,y,z;
+  DOUBLE   anorm,c,f,g,h,s,scale,x,y,z;
 
   g = scale = anorm = 0.0; // Householder reduction to bidiagonal form.
   for (i=0;i<n;i++) {
@@ -671,8 +671,8 @@ void Matrix::svd(Matrix &U2,Matrix &W,Matrix &V) {
       flag = 1;
       for (l=k;l>=0;l--) { // Test for splitting.
         nm = l-1;
-        if ((FLOAT)(fabs(rv1[l])+anorm) == anorm) { flag = 0; break; }
-        if ((FLOAT)(fabs( w[nm])+anorm) == anorm) { break; }
+        if ((DOUBLE)(fabs(rv1[l])+anorm) == anorm) { flag = 0; break; }
+        if ((DOUBLE)(fabs( w[nm])+anorm) == anorm) { break; }
       }
       if (flag) {
         c = 0.0; // Cancellation of rv1[l], if l > 1.
@@ -680,7 +680,7 @@ void Matrix::svd(Matrix &U2,Matrix &W,Matrix &V) {
         for (i=l;i<=k;i++) {
           f = s*rv1[i];
           rv1[i] = c*rv1[i];
-          if ((FLOAT)(fabs(f)+anorm) == anorm) break;
+          if ((DOUBLE)(fabs(f)+anorm) == anorm) break;
           g = w[i];
           h = pythag(f,g);
           w[i] = h;
@@ -760,9 +760,9 @@ void Matrix::svd(Matrix &U2,Matrix &W,Matrix &V) {
   // by decreasing magnitude. Also, signs of corresponding columns are
   // flipped so as to maximize the number of positive elements.
   int32_t s2,inc=1;
-  FLOAT   sw;
-  FLOAT* su = (FLOAT*)malloc(m*sizeof(FLOAT));
-  FLOAT* sv = (FLOAT*)malloc(n*sizeof(FLOAT));
+  DOUBLE   sw;
+  DOUBLE* su = (DOUBLE*)malloc(m*sizeof(DOUBLE));
+  DOUBLE* sv = (DOUBLE*)malloc(n*sizeof(DOUBLE));
   do { inc *= 3; inc++; } while (inc <= n);
   do {
     inc /= 3;
@@ -829,8 +829,8 @@ void Matrix::allocateMemory (const int32_t m_,const int32_t n_) {
     val = 0;
     return;
   }
-  val    = (FLOAT**)malloc(m*sizeof(FLOAT*));
-  val[0] = (FLOAT*)calloc(m*n,sizeof(FLOAT));
+  val    = (DOUBLE**)malloc(m*sizeof(DOUBLE*));
+  val[0] = (DOUBLE*)calloc(m*n,sizeof(DOUBLE));
   for(int32_t i=1; i<m; i++)
     val[i] = val[i-1]+n;
 }
@@ -842,8 +842,8 @@ void Matrix::releaseMemory () {
   }
 }
 
-FLOAT Matrix::pythag(FLOAT a,FLOAT b) {
-  FLOAT absa,absb;
+DOUBLE Matrix::pythag(DOUBLE a,DOUBLE b) {
+  DOUBLE absa,absb;
   absa = fabs(a);
   absb = fabs(b);
   if (absa > absb)
